Removed `Engine::{run,dump}File`, which were identical in the two engines.
`Joe::{run,dump}File` now contains the relevant file-reading code, and calls
`Joe::{run,dump}`, which delegates to `Engine::{run,dump}`.
