+++ lib/src/main/java/com/wjduquette/joe/walker/WalkerEngine.java
@@ -2,10 +2,6 @@ package com.wjduquette.joe.walker;
 
 import com.wjduquette.joe.*;
 
-import java.io.IOException;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -47,26 +43,6 @@ public class WalkerEngine implements Engine {
     @Override
     public void setVar(String name, Object value) {
         interpreter.globals().setVar(name, value);
-    }
-
-    /**
-     * Reads the given file and executes its content as a script.
-     * @param scriptPath The file's path
-     * @return The script's result
-     * @throws IOException if the file cannot be read.
-     * @throws SyntaxError if the script could not be compiled.
-     * @throws JoeError on all runtime errors.
-     */
-    @SuppressWarnings("UnusedReturnValue")
-    @Override
-    public Object runFile(String scriptPath)
-        throws IOException, SyntaxError, JoeError
-    {
-        var path = Paths.get(scriptPath);
-        byte[] bytes = Files.readAllBytes(path);
-        var script = new String(bytes, Charset.defaultCharset());
-
-        return run(path.getFileName().toString(), script);
     }
 
     /**
@@ -159,22 +135,6 @@ public class WalkerEngine implements Engine {
         } else {
             throw joe.expected("callable", callee);
         }
-    }
-
-    /**
-     * Compiles the given file and returns a compilation dump.
-     * @param scriptPath The file's path
-     * @return The script's result
-     * @throws IOException if the file cannot be read.
-     * @throws SyntaxError if the script could not be compiled.
-     */
-    @SuppressWarnings("UnusedReturnValue")
-    public String dumpFile(String scriptPath) throws IOException, SyntaxError {
-        var path = Paths.get(scriptPath);
-        byte[] bytes = Files.readAllBytes(path);
-        var script = new String(bytes, Charset.defaultCharset());
-
-        return dump(path.getFileName().toString(), script);
     }
 
     @Override
